requirements.txt
pandas>=2.2.0  # Data manipulation
pyarrow>=15.0.0  # Parquet file reading
geopandas>=0.14.0  # Geospatial data handling
minio>=7.2.0  # MinIO client for CSV storage/retrieval
pydantic>=2.8.0  # Settings validation
python-dotenv>=1.0.0  # Environment variable loading
structlog>=24.4.0  # Structured logging
tenacity>=8.5.0  # Retry logic 
tqdm>=4.66.0  # Progress bars
networkx>=3.3  # Graph-based route optimization
shapely>=2.0.0  # Geometry operations
duckdb>=1.0.0  # In-memory SQL for batch processing
typer>=0.12.0  # CLI interface
pytest>=8.3.0  # Unit testing 
sqlite3  # Built-in Python library
black>=24.8.0  # Code formatting